Event Handlers are all good now
The Inventory System Is pretty much done now, any other changes will
happen later on over time as I see fit.
Removing an object is a bit funky but is functional.
Equipping and un equipping is good now and all the event handlers for
the Inventory System are in place for use with UI updates.