Cleanup
Moved "specialized" entity code into its own example class and made the
internal Entity class abstract.
An Entity is now simply a class that has a statContainer in it, anything
further would be a specialization that is project specific.
Probably want to make Stats and Level (the properties) protected
(instead of public).

Removed all specific StatName references out of the Systems folder.
Hopefully, nothing will break if anyone takes just the Systems folder.
(We should figure out a way to subtree it...)

Note: InventoryManager was also made abstract and one of the  properties
unimplemented (though it probably shouldn't have been implemented as a
base for the system)